requirements.txt
falcon==1.1.0
Cython==0.25.2
colander==1.3.2
pymongo==3.4
mongoengine==0.11.0
gunicorn==19.6.0
PyJWT==1.4.2

